+++ data-raw/psychsci.R
@@ -1,0 +1,5 @@
+## code to prepare `psychsci` dataset goes here
+
+psychsci <- read_grobid("vignettes/xml")
+
+usethis::use_data(psychsci, overwrite = TRUE)